[Refactor] Extract DevBuddy to another file
