+++ src/extension.ts
@@ -1,6 +1,5 @@
-/* eslint-disable @typescript-eslint/naming-convention */
 import * as vscode from 'vscode';
-import { Configuration, OpenAIApi } from "openai";
+import { DevBuddy } from './dev-buddy';
 
 export function activate(context: vscode.ExtensionContext) {
 	console.log('Congratulations, your extension "dev-buddy-openai" is now active!');
@@ -10,8 +9,7 @@ export function activate(context: vscode.ExtensionContext) {
 		vscode.window.showErrorMessage('The OPENAI_API_KEY environment variable is not set');
 		return;
 	}
-	const configuration = new Configuration({ apiKey: openaiApiKey });
-	const openai = new OpenAIApi(configuration);
+	const devBuddy = new DevBuddy(openaiApiKey);
 
 	let disposable = vscode.commands.registerCommand('dev-buddy-openai.completeComments', async () => {
 		const editor = vscode.window.activeTextEditor;
@@ -22,33 +20,17 @@ export function activate(context: vscode.ExtensionContext) {
 		const languageId = editor.document.languageId;
 		const selection = editor.selection;
 		const selectedText = editor.document.getText(selection);
-		const requestText = [
-			'Replace all the "TODO" and "FIXME" comments from the following code',
-			'for the code that actually does what the comment expects',
-			'and remove the entire comment after the "TODO" or "FIXME":',
-			'```' + languageId,
-			selectedText,
-			'```'
-		].join("\n");
 
-		try {
-			const completion = (await openai.createCompletion({
-				model: "text-davinci-003",
-				prompt: requestText,
-				temperature: 0.7,
-				max_tokens: 256,
-				top_p: 1,
-				frequency_penalty: 0,
-				presence_penalty: 0,
-			})).data;
-			const newText = completion.choices[0].text?.trim() || selectedText;
-
+		const functionSuccess = (newText: string) => {
 			editor.edit(editBuilder => {
 				editBuilder.replace(selection, newText);
 			});
-		} catch (error: any) {
-			vscode.window.showErrorMessage(`An error occurred while completing the code: ${error.message}`);
-		}
+		};
+		const functionFailure = (errorMessage: string) => {
+			vscode.window.showErrorMessage(`An error occurred while completing the code: ${errorMessage}`);
+		};
+
+		devBuddy.completeComments(selectedText, languageId, functionSuccess, functionFailure);
 	});
 
 	context.subscriptions.push(disposable);
